seshat: Move more stuff out of the types.rs file.
